requirements.txt
customtkinter
svglib
reportlab
Pillow